+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Current (in progress)
 
+- Allow temporal coverage with only a start date [#3192](https://github.com/opendatateam/udata/pull/3192)
 - The `extras` column in the resource catalog is now dumped as json [#3272](https://github.com/opendatateam/udata/pull/3272) and [#3273](https://github.com/opendatateam/udata/pull/3273)
 
 ## 10.1.0 (2025-02-20)